+++ packages/native-ui/docs/components/AllComponents.web.tsx
@@ -24,8 +24,9 @@ import {
   useColorMode,
   Card,
   Switch,
-} from '../../src';
-import { StyleSheet } from 'react-native-unistyles';
+  ToggleButton,
+  ToggleButtonGroup,
+} from '@utilitywarehouse/native-ui';
 import {
   Actionsheet,
   Box,
@@ -33,7 +34,9 @@ import {
   CarouselItem,
   CarouselItems,
   CarouselPagination,
-} from '../../src/lab';
+  Accordion,
+  AccordionItem,
+} from '@utilitywarehouse/native-ui/lab';
 import {
   ElectricityMediumIcon,
   MobileMediumIcon,
@@ -69,12 +72,28 @@ const AllComponents: React.FC = () => {
   const toggleActionsheet = () => setShowActionsheet(!showActionsheet);
   const [switchEnabled, setSwitchEnabled] = React.useState(false);
   const toggleSwitch = () => setSwitchEnabled(!switchEnabled);
+  const [toggleButtonValue, setToggleButtonValue] = React.useState('day');
   const colorMode = useColorMode();
   const isDark = colorMode === 'dark';
   return (
     <div className="sb-unstyled">
       <ScrollView contentContainerStyle={styles.container}>
         <HStack wrap space="md">
+          <ComponentWrapper
+            name="Accordion (Lab)"
+            link="/?path=/docs/native-ui-components-lab-accordion--docs"
+          >
+            <Center flex={1}>
+              <Accordion type="single">
+                <AccordionItem title="Accordion Item 1">
+                  <Text>Accordion Item 1 Content</Text>
+                </AccordionItem>
+                <AccordionItem title="Accordion Item 2">
+                  <Text>Accordion Item 2 Content</Text>
+                </AccordionItem>
+              </Accordion>
+            </Center>
+          </ComponentWrapper>
           <ComponentWrapper
             name="Actionsheet (Lab)"
             link="/?path=/docs/components-actionsheet-lab--docs"
@@ -257,7 +276,19 @@ const AllComponents: React.FC = () => {
               <Text>This is some Text</Text>
             </Center>
           </ComponentWrapper>
-          <ComponentWrapper name="VStack" link="/?path=/docs/components-vstack--docs">
+          <ComponentWrapper
+            name="Toggle Button"
+            link="/?path=/docs/native-ui-components-togglebutton--docs"
+          >
+            <Center flex={1}>
+              <ToggleButtonGroup value={toggleButtonValue} onChange={setToggleButtonValue}>
+                <ToggleButton value="day">Day</ToggleButton>
+                <ToggleButton value="week">Week</ToggleButton>
+                <ToggleButton value="month">Month</ToggleButton>
+              </ToggleButtonGroup>
+            </Center>
+          </ComponentWrapper>
+          <ComponentWrapper name="VStack" link="/?path=/docs/native-ui-components-vstack--docs">
             <Center flex={1}>
               <VStack space="md">
                 <Box w={20} h={20} bg="$cyan300" />