+++ packages/native-ui/src/components/ToggleButton/ToggleButtonGroup.props.ts
@@ -1,0 +1,16 @@
+import { ViewProps } from 'react-native';
+
+export interface ToggleButtonGroupProps<T = string | number | boolean>
+  extends Omit<ViewProps, 'onChange'> {
+  /**
+   * The currently selected value
+   */
+  value: T;
+
+  /**
+   * Callback fired when the value changes
+   */
+  onChange?: (value: T) => void;
+  disabled?: boolean;
+  size?: 'small' | 'base';
+}