+++ packages/native-ui/src/lab/Accordion/AccordionItem.context.tsx
@@ -1,0 +1,11 @@
+import { createContext, useContext } from 'react';
+
+const AccordionItemContext = createContext<{
+  noPadding?: boolean;
+  disabled?: boolean;
+  divider?: boolean;
+}>({});
+
+export const useAccordionItemContext = () => useContext(AccordionItemContext);
+
+export default AccordionItemContext;