+++ packages/native-ui/src/lab/Accordion/index.ts
@@ -1,0 +1,10 @@
+export {
+  default as Accordion,
+  AccordionItem,
+  AccordionHeader,
+  AccordionTrigger,
+  AccordionContent,
+  AccordionContentText,
+  AccordionIcon,
+  AccordionTitleText,
+} from './Accordion';