+++ packages/react/.eslintrc.cjs
@@ -1,8 +1,0 @@
-/* eslint-env node */
-module.exports = {
-  extends: ['../../.eslintrc.js', 'plugin:storybook/recommended'],
-  parserOptions: {
-    project: './tsconfig.json',
-    tsconfigRootDir: __dirname,
-  },
-};