+++ packages/react/src/components/index.ts
@@ -1,5 +1,0 @@
-export { Box } from './Box/Box';
-export type { BoxProps } from './Box/Box.props';
-
-export { BodyText } from './BodyText/BodyText';
-export type { BodyTextProps } from './BodyText/BodyText.props';